# api/openapi.yaml
openapi: "3.0.3"
info:
  title: ShiftManager API
  version: "1.0.0"
  description: REST API for the ShiftManager club volunteer-hours system.
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"

    # ── Member ────────────────────────────────────────────────────────────────
...
    EventStatus:
      type: string
      enum: [draft, published, cancelled, completed]
    EventVisibility:
      type: string
      enum: [public, members, hidden]
    Event:
      type: object
      required: [id, name, status, visibility]
      properties:
        id:          { $ref: '#/components/schemas/UUID' }
        name:        { type: string }
        description: { type: string }
        location:    { type: string }
        category:    { type: string }
        status:      { $ref: '#/components/schemas/EventStatus' }
        visibility:  { $ref: '#/components/schemas/EventVisibility' }
        createdAt:   { type: string, format: date-time }
        updatedAt:   { type: string, format: date-time }
    EventWrite:
      type: object
      required: [name, status, visibility]
      properties:
        name:        { type: string }
        description: { type: string }
        location:    { type: string }
        category:    { type: string }
        status:      { $ref: '#/components/schemas/EventStatus' }
        visibility:  { $ref: '#/components/schemas/EventVisibility' }

    # ── Shift ─────────────────────────────────────────────────────────────────
    Shift:
      type: object
      required: [id, eventId, name, startAt, endAt, minHelpers, maxHelpers]
      properties:
        id:                    { $ref: '#/components/schemas/UUID' }
        eventId:               { $ref: '#/components/schemas/UUID' }
        name:                  { type: string }
        startAt:               { type: string, format: date-time }
        endAt:                 { type: string, format: date-time }
        minHelpers:            { type: integer }
        maxHelpers:            { type: integer }
        requiredQualification: { type: string }
...
    ShiftOccupancy:
      type: object
      required: [registered, confirmed, min, max]
      properties:
        registered: { type: integer }
        confirmed:  { type: integer }
        min:        { type: integer }
        max:        { type: integer }
    Signup:
      type: object
      required: [id, memberId, state]
      properties:
        id:       { $ref: '#/components/schemas/UUID' }
        memberId: { $ref: '#/components/schemas/UUID' }
        state:    { type: string }
        comment:  { type: string }
    ShiftDay:
      type: object
      required: [date, shifts]
      properties:
        date:   { type: string, format: date }
        shifts:
          type: array
          items: { $ref: '#/components/schemas/ShiftWithOccupancy' }
    ShiftWithOccupancy:
      allOf:
        - $ref: '#/components/schemas/Shift'
        - type: object
          properties:
            occupancy: { $ref: '#/components/schemas/ShiftOccupancy' }
            signups:
              type: array
              items: { $ref: '#/components/schemas/Signup' }
    EventTimeline:
      type: object
      required: [event, days]
      properties:
        event: { $ref: '#/components/schemas/Event' }
        days:
          type: array
          items: { $ref: '#/components/schemas/ShiftDay' }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Insufficient role
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
security:
  - bearerAuth: []
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/EventStatus' } }
        - { name: visibility, in: query, schema: { $ref: '#/components/schemas/EventVisibility' } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Event' } }
        "401": { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event (Veranstaltungsleiter+)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventWrite' }
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Event' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /events/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event with full shift timeline
      responses:
        "200":
          description: Event timeline
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventTimeline' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event (Veranstaltungsleiter+)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventWrite' }
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Event' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event (Veranstaltungsleiter+, draft/cancelled only)
      responses:
        "204": { description: Deleted }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /events/{id}/publish:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    post:
      operationId: publishEvent
      tags: [events]
      summary: Publish a draft event (Veranstaltungsleiter+)
      responses:
        "200":
          description: Published event
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Event' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /events/{id}/timeline:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    get:
      operationId: getEventTimeline
      tags: [events]
      summary: Get just the shift timeline for an event
      responses:
        "200":
          description: Event timeline
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventTimeline' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
...
  /kiosk/events:
    get:
      operationId: listKioskEvents
      tags: [kiosk]
      summary: List all published events (no auth)
      security: []
      responses:
        "200":
          description: Published events
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Event' } }
        "403": { description: Kiosk locked }
  /kiosk/events/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    get:
      operationId: getKioskEventTimeline
      tags: [kiosk]
      summary: Get published event timeline (no auth)
      security: []
      responses:
        "200":
          description: Event timeline
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventTimeline' }
        "403": { description: Kiosk locked }
        "404": { $ref: '#/components/responses/NotFound' }
...